refactor(InputView): move const val into InputView
