+++ src/main/kotlin/lotto/Main.kt
@@ -1,23 +1,20 @@
 import lotto.domain.LottoItems
-import lotto.domain.LottoNumber
-import lotto.input.InputView
+import lotto.view.InputView
 
 fun main() {
-    val inputMoney = InputView.inputMoney("구입금액을 입력해 주세요.")
+    val inputMoney = InputView.inputMoney()
     val buyCount = inputMoney / 1000
     println("${buyCount}개를 구매했습니다.")
 
     val lottoItems = LottoItems(buyCount)
     val lottos = lottoItems.getLottoItems()
 
-    // InputView.inputWinningNumber("지난 주 당첨 번호를 입력해 주세요.")
-    val winningNumbers = List(6) { i -> LottoNumber(i + 1) }
+    val winningNumbers = InputView.inputWinningNumber()
 
-    println(lottos)
-
-    val a = lottoItems.getWinLottos(winningNumbers)
+    println(lottos.joinToString(""))
 
     println("당첨 통계")
     println("---------")
-    println(a.joinToString(""))
+    val result = lottoItems.getWinLottos(winningNumbers)
+    println(result.joinToString(""))
 }
