+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,17 +1,22 @@
-package lotto.input
+package lotto.view
 
-import java.lang.IllegalArgumentException
+import lotto.domain.value.LottoNumber
 
 object InputView {
-    fun inputMoney(value: String): Int {
-        println(value)
-        return readLine()?.toInt() ?: throw IllegalArgumentException("숫자만 넣어 주세요")
+    const val INPUT_MONEY_MESSAGE = "구입금액을 입력해 주세요."
+    const val INPUT_NOT_INTEGER_ERROR_MESSAGE = "숫자만 넣어 주세요"
+    const val INPUT_WINNING_NUMBER_MESSAGE = "지난 주 당첨 번호를 입력해 주세요."
+    const val INPUT_NOT_VALID_ERROR_MESSAGE = "콤마(,)로 연결된 6자리 숫자를 입력하세요."
+
+    fun inputMoney(): Int {
+        println(INPUT_MONEY_MESSAGE)
+        return readLine()?.toInt() ?: throw IllegalArgumentException(INPUT_NOT_INTEGER_ERROR_MESSAGE)
     }
 
-    fun inputWinningNumber(value: String): List<Int> {
-        println(value)
+    fun inputWinningNumber(): List<LottoNumber> {
+        println(INPUT_WINNING_NUMBER_MESSAGE)
         return readLine()?.split(",")?.map {
-            it.toInt()
-        } ?: throw IllegalArgumentException("콤마(,)로 연결된 6자리 숫자를 입력하세요.")
+            LottoNumber(it.toInt())
+        } ?: throw IllegalArgumentException(INPUT_NOT_VALID_ERROR_MESSAGE)
     }
 }
